+++ src/app/(navRoute)/edit-feed/[id]/page.tsx
@@ -1,12 +1,11 @@
 import { Metadata } from 'next';
 
-// import EditFeed from '@/containers/editFeed/EditFeed';
+import EditFeed from '@/containers/editFeed/EditFeed';
 
 export const metadata: Metadata = {
   title: 'Edit feed | MOMOO',
 };
 
 export default function page() {
-  return <div>준비중</div>;
-  //   return <EditFeed />;
+  return <EditFeed />;
 }
