+++ src/components/FeedItem/FeedItem.tsx
@@ -1,44 +1,35 @@
 import Image from 'next/image';
-import { useParams } from 'next/navigation';
-// import { ForwardedRef, forwardRef, useEffect, useState } from 'react';
-import { ForwardedRef, forwardRef, useEffect, useMemo, useState } from 'react';
+import { useParams, useRouter } from 'next/navigation';
+import { ForwardedRef, forwardRef, useMemo, useState } from 'react';
 
-import { DocumentData } from '@firebase/firestore';
-// import { useSelector } from 'react-redux';
 import { Timestamp } from 'firebase/firestore';
 
 import Carousel from '@/components/Carousel/Carousel';
+import EditFeedModal from '@/components/EditFeed/EditFeedModal';
 import StyledFeedItem from '@/components/FeedItem/StyledFeedItem';
 import ChangeAlbumModal from '@/components/Modal/ChangeAlbumModal/ChangeAlbumModal';
 import DeleteFeedModal from '@/components/Modal/DeleteFeedModal';
 import FeedMoreModal from '@/components/Modal/FeedMoreModal';
 import useAuthState from '@/hooks/auth/useAuthState';
-// import useGetFeedData from '@/hooks/useGetFeedData';
+import useWindowWidth from '@/hooks/useWindowWidth';
 
-// import { ReduxState } from '@/modules/model';
+import type { Feed } from '@/types/feed';
 
-function FeedItem(
-  { feed }: { feed: DocumentData },
-  ref: ForwardedRef<HTMLLIElement>,
-) {
+function FeedItem({ feed }: { feed: Feed }, ref: ForwardedRef<HTMLLIElement>) {
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [deleteModalOpen, setDeleteModalOpen] = useState(false);
   const [changeAlbumModalOpen, setChangeAlbumModalOpen] = useState(false);
-  const [feedData, setFeedData] = useState<DocumentData | null>(feed);
+  const [feedData, setFeedData] = useState<Feed | null>(feed);
+  const [isEditModalOpen, setIsEditModalOpen] = useState(false);
+
   const { uid } = useParams<{
     uid: string;
     albumName: string;
   }>();
-
-  useEffect(() => {
-    setFeedData((prev) => prev);
-  }, []);
+  const router = useRouter();
 
   const { user } = useAuthState();
-  // const isEditFeedModalOpen = useSelector(
-  //   (state: ReduxState) => state.editFeedModal.isEditFeedModalOpen,
-  // );
-  // const getFeedData = useGetFeedData();
+  const windowWidth = useWindowWidth();
 
   const getFormattedDateFromTimestamp = (timestamp: Timestamp) => {
     const date = new Date(
@@ -54,32 +45,6 @@ function FeedItem(
     () => feedData && getFormattedDateFromTimestamp(feedData.timestamp),
     [feedData],
   );
-
-  // useEffect(() => {
-  //   // 피드 수정 후 리렌더링
-  //   if (isEditFeedModalOpen) {
-  //     return;
-  //   }
-
-  //   const setData = async () => {
-  //     const feedData = await getFeedData(id, uid);
-
-  //     if (feedData) {
-  //       setFeedData(feedData);
-
-  //       const date = new Date(feedData.timestamp.toDate());
-  //       const time = new Date(date.setHours(date.getHours() + 9))
-  //         .toISOString()
-  //         .slice(0, 10);
-
-  //       setTime(time);
-  //     } else {
-  //       setFeedData(null);
-  //     }
-  //   };
-
-  //   setData();
-  // }, [isEditFeedModalOpen]);
 
   const handleSeeMoreClick = () => {
     setIsModalOpen(true);
@@ -99,11 +64,21 @@ function FeedItem(
     setIsModalOpen(false);
   };
 
+  if (!feedData) {
+    return <></>;
+  }
+
+  const goToEditFeed = () => {
+    if (windowWidth && windowWidth > 430) {
+      setIsEditModalOpen(true);
+    } else {
+      router.push(`/edit-feed/${feedData.id}`);
+    }
+  };
+
   return (
     <>
-      {!feedData ? (
-        <></>
-      ) : (
+      {feedData && (
         <StyledFeedItem ref={ref}>
           <Carousel imgUrlList={feedData.imageUrl}></Carousel>
           <section className="contentsSection">
@@ -188,10 +163,12 @@ function FeedItem(
               </>
             )}
           </div>
+
           {isModalOpen && (
             <FeedMoreModal
               setDeleteModalOpen={setDeleteModalOpen}
               setChangeAlbumModalOpen={setChangeAlbumModalOpen}
+              goToEditFeed={goToEditFeed}
               closeModal={closeMoreModal}
             />
           )}
@@ -203,6 +180,13 @@ function FeedItem(
           )}
           {changeAlbumModalOpen && (
             <ChangeAlbumModal onClose={handleChangeAlbumModal} />
+          )}
+          {isEditModalOpen && (
+            <EditFeedModal
+              id={feedData.id}
+              setFeedData={setFeedData}
+              setIsModalOpen={setIsEditModalOpen}
+            />
           )}
         </StyledFeedItem>
       )}
